Move drawing controls component to own file
